# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rare Gems API Docs",
        "description": "Rare Gems API Docs",
        "contact": {
            "email": "rafiumoshoodolakunle@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://app.raregems.ng",
            "description": "API Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/packages": {
            "get": {
                "tags": [
                    "Package"
                ],
                "summary": "Get Packages",
                "operationId": "get packages",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "should be enabled or disabled",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Unprocessed Entity",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/packages/{id}/show": {
            "get": {
                "tags": [
                    "Package"
                ],
                "summary": "Show Package",
                "operationId": "show package",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}